+++ board/khadas/rk3588s-khadas-edge2/kedge2.c
@@ -7,6 +7,10 @@
 #include <common.h>
 #include <dwc3-uboot.h>
 #include <usb.h>
+#include <i2c.h>
+#include <dm.h>
+
+#define TP_I2C_BUS_NUM 6
 
 DECLARE_GLOBAL_DATA_PTR;
 
@@ -41,4 +45,46 @@ int rk_board_late_init(void)
 	run_command("gpio set 138; gpio clear 139; gpio clear 140", 0);
 
 	return 0;
+}
+
+int rk_board_init(void)
+{
+	int ret = 0;
+	int res = 0;
+	struct udevice *bus;
+	struct udevice *dev;
+	uchar linebuf[1];
+
+	run_command("gpio set 130", 0);//GPIO4_A2 vcc 5v
+
+	uclass_get_device_by_seq(UCLASS_I2C, TP_I2C_BUS_NUM, &bus);
+	ret = i2c_get_chip(bus, 0x38, 1, &dev);
+	if (!ret) {
+		res = dm_i2c_read(dev, 0xA8, linebuf, 1);
+		if (!res) {
+			printf("TP05 id=0x%x\n", linebuf[0]);
+			if (linebuf[0] == 0x51){//old ts050
+				env_set("lcd_panel","ts050");
+			} else if (linebuf[0] == 0x79) {//new ts050
+				env_set("lcd_panel","newts050");
+			}
+		}
+	}
+	if (ret || res) {
+		ret = i2c_get_chip(bus, 0x14, 1, &dev);
+		if (!ret) {
+			res = dm_i2c_read(dev, 0x9e, linebuf, 1);
+			if (!res) {
+				printf("TP10 id=0x%x\n", linebuf[0]);
+				if (linebuf[0] == 0x00) {//TS101
+					env_set("lcd_panel","ts101");
+				}
+			} else {
+				env_set("lcd_panel","null");
+			}
+		}
+	}
+
+	return 0;
+
 }
